+++ src/ic_registry/projection/mod.rs
@@ -1,0 +1,11 @@
+mod data_center;
+mod node;
+mod node_operator;
+mod node_provider;
+
+pub(super) use data_center::data_center_list_from_inventory;
+pub(super) use node::node_list_from_inventory;
+pub(super) use node_operator::node_operator_list_from_inventory;
+#[cfg(test)]
+pub(super) use node_provider::node_provider_from_governance;
+pub(super) use node_provider::node_provider_list_from_response;